+++ Domitransp/app/models/company.rb
@@ -1,7 +1,5 @@
 class Company < ApplicationRecord
     has_many :users
     mount_uploader :rut, RutUploader
-    def user_count
-        users.count
-    end
+    
 end
