+++ Domitransp/app/models/user.rb
@@ -1,7 +1,8 @@
 class User < ApplicationRecord
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
-  belongs_to :company
+  belongs_to :company, , counter_cache: true
+  
   devise :database_authenticatable, :registerable,
-        :recoverable, :rememberable, :validatable
+    :recoverable, :rememberable, :validatable
 end
