+++ Domitransp/app/views/companies/_form.html.erb
@@ -1,12 +1,8 @@
 <%= simple_form_for(@company) do |f| %>
     <%= f.error_notification %>
-    <%= f.error_notification message: f.object.errors[:base].to_sentence if f.object.errors[:base].present? %>
-    
+    <%= f.error_notification message: f.object.errors[:base].to_sentence if f.object.errors[:base].present? %> 
     <%= f.input :nit %>
     <%= f.input :razon_social %>
-    
-
-
     <%= f.input :rut, label: false %>
     <%= f.submit %>
 <% end %>
