+++ Domitransp/app/views/companies/index.html.erb
@@ -21,14 +21,13 @@
                     <a class="collapse-link">
                         <i class="fa fa-chevron-up"></i>
                     </a>
-                    
-                    
-                    
                 </div>
             </div>
             <div class="ibox-content">
                 <div class="row">
-                    <%= render 'filter' %>
+                    <div class="col-lg-8">
+                        <%= render 'filter' %>
+                    </div>
                 </div>
