+++ Domitransp/app/views/companies/new.html.erb
@@ -22,7 +22,6 @@
             </div>
         <div class="ibox-content">
             <%= render 'form', company: @company %>
-
         </div>
     </div>
 </div>
